Add ability to hide password images
This disables loading of favicon images associated with password entries
and hides any images that are already loaded, using the generic icon
instead.

The key benefit to this option is to prevent passforios revealing that a
given device has a password in its store, which could be gleaned from
the fact that favicons are being loaded in this manner.