Localise extensions by copying bundle resources
Previously, the Pass extensions (for autofill and for the sharing sheet)
didn't see their strings being localised. This means that in some cases,
English was the only thing shown and in others strings would look like
`#EnterPasscode#` on the passcode lock screen.  This was because their
build phases didn't reference the localised string files in the `Copy
Bundle Resources` phase.

This change ensures these localisation files are copied and thus the
extensions are now localised for both English and German (so far).