Move persistence controller setup into the init
App extensions access the persistent storage. From the refactor, the
persistent storage "setup" is no longer called which breaks the
extension and the pass auto fill.

This commit moves the persistent storage setup into the init for the
shared storage, which should be called once for the first read access or
in preparation for a write.

As a result, extensions should now work as expected.